Reorganize Tile and Text File Structure
