+++ src/components/common/tiles/AddressTile.js
@@ -7,20 +7,20 @@ import { AppState, TouchableOpacity, View } from 'react-native'
 import FontAwesome from 'react-native-vector-icons/FontAwesome'
 import FontAwesome5 from 'react-native-vector-icons/FontAwesome5'
 
-import { launchBitPay } from '../../actions/BitPayActions.js'
-import { addressWarnings } from '../../actions/ScanActions.js'
-import s from '../../locales/strings.js'
-import { checkPubAddress } from '../../modules/FioAddress/util'
-import { BitPayError } from '../../types/BitPayError.js'
-import { forwardRef } from '../../types/reactHooks.js'
-import { connect } from '../../types/reactRedux.js'
-import { type GuiMakeSpendInfo } from '../../types/types.js'
-import { parseDeepLink } from '../../util/DeepLinkParser.js'
-import { AddressModal } from '../modals/AddressModal'
-import { ScanModal } from '../modals/ScanModal.js'
-import { Airship, showError } from '../services/AirshipInstance'
-import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
-import { EdgeText } from './EdgeText'
+import { launchBitPay } from '../../../actions/BitPayActions.js'
+import { addressWarnings } from '../../../actions/ScanActions.js'
+import s from '../../../locales/strings.js'
+import { checkPubAddress } from '../../../modules/FioAddress/util'
+import { BitPayError } from '../../../types/BitPayError.js'
+import { forwardRef } from '../../../types/reactHooks.js'
+import { connect } from '../../../types/reactRedux.js'
+import { type GuiMakeSpendInfo } from '../../../types/types.js'
+import { parseDeepLink } from '../../../util/DeepLinkParser.js'
+import { AddressModal } from '../../modals/AddressModal'
+import { ScanModal } from '../../modals/ScanModal.js'
+import { Airship, showError } from '../../services/AirshipInstance'
+import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../../services/ThemeContext.js'
+import { EdgeText } from '../../themed/EdgeText'
 import { Tile } from './Tile.js'
 
 type OwnProps = {
