+++ src/components/common/tiles/EditableAmountTile.js
@@ -3,13 +3,13 @@ import { div, toFixed } from 'biggystring'
 import { type EdgeCurrencyWallet, type EdgeDenomination } from 'edge-core-js'
 import * as React from 'react'
 
-import s from '../../locales/strings.js'
-import { convertCurrencyFromExchangeRates } from '../../selectors/WalletSelectors.js'
-import { type GuiExchangeRates } from '../../types/types.js'
-import { getWalletFiat } from '../../util/CurrencyWalletHelpers.js'
-import { DECIMAL_PRECISION, getDenomFromIsoCode, zeroString } from '../../util/utils'
-import { type Theme, cacheStyles, useTheme } from '../services/ThemeContext.js'
-import { EdgeText } from './EdgeText'
+import s from '../../../locales/strings.js'
+import { convertCurrencyFromExchangeRates } from '../../../selectors/WalletSelectors.js'
+import { type GuiExchangeRates } from '../../../types/types.js'
+import { getWalletFiat } from '../../../util/CurrencyWalletHelpers.js'
+import { DECIMAL_PRECISION, getDenomFromIsoCode, zeroString } from '../../../util/utils'
+import { type Theme, cacheStyles, useTheme } from '../../services/ThemeContext.js'
+import { EdgeText } from '../../themed/EdgeText'
 import { Tile } from './Tile.js'
 
 type Props = {
