+++ src/components/common/tiles/Tile.js
@@ -6,10 +6,10 @@ import { ActivityIndicator, TouchableWithoutFeedback, View } from 'react-native'
 import FontAwesomeIcon from 'react-native-vector-icons/FontAwesome'
 import SimpleLineIcons from 'react-native-vector-icons/SimpleLineIcons'
 
-import s from '../../locales/strings.js'
-import { showToast } from '../services/AirshipInstance.js'
-import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
-import { EdgeText } from './EdgeText.js'
+import s from '../../../locales/strings.js'
+import { showToast } from '../../services/AirshipInstance.js'
+import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../../services/ThemeContext.js'
+import { EdgeText } from '../../themed/EdgeText'
 
 const textHeights = {
   small: 2,
