+++ src/components/modals/TransactionAdvanceDetails.js
@@ -7,11 +7,11 @@ import { type AirshipBridge } from 'react-native-airship'
 import SafariView from 'react-native-safari-view'
 
 import s from '../../locales/strings.js'
+import { Tile } from '../common/tiles/Tile'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 import { EdgeText } from '../themed/EdgeText.js'
 import { ModalCloseArrow } from '../themed/ModalParts.js'
 import { ThemedModal } from '../themed/ThemedModal.js'
-import { Tile } from '../themed/Tile.js'
 
 const localizedFeeText = {
   satPerVByte: s.strings.transaction_details_advance_details_satpervbyte,
