+++ src/components/modals/WcSmartContractModal.js
@@ -15,13 +15,13 @@ import { getDenominationFromCurrencyInfo } from '../../selectors/DenominationSel
 import { useSelector } from '../../types/reactRedux.js'
 import { getCurrencyIconUris } from '../../util/CdnUris'
 import { hexToDecimal, isHex, removeHexPrefix, zeroString } from '../../util/utils.js'
+import { CryptoFiatAmountTile } from '../common/tiles/CryptoFiatAmountTile'
+import { FiatAmountTile } from '../common/tiles/FiatAmountTile'
+import { IconTile } from '../common/tiles/IconTile'
 import { Airship, showError } from '../services/AirshipInstance.js'
 import { type Theme, cacheStyles, useTheme } from '../services/ThemeContext.js'
 import { Alert } from '../themed/Alert'
-import { CryptoFiatAmountTile } from '../themed/CryptoFiatAmountTile.js'
 import { EdgeText } from '../themed/EdgeText'
-import { FiatAmountTile } from '../themed/FiatAmountTile.js'
-import { IconTile } from '../themed/IconTile'
 import { ModalCloseArrow, ModalTitle } from '../themed/ModalParts.js'
 import { ThemedModal } from '../themed/ThemedModal.js'
