+++ src/components/scenes/FioAddressSettingsScene.js
@@ -10,12 +10,12 @@ import { addBundledTxs, getAddBundledTxsFee, getTransferFee } from '../../module
 import { connect } from '../../types/reactRedux.js'
 import { type NavigationProp, type RouteProp } from '../../types/routerTypes.js'
 import { SceneWrapper } from '../common/SceneWrapper'
+import { Tile } from '../common/tiles/Tile'
 import { ButtonsModal } from '../modals/ButtonsModal'
 import { Airship, showError, showToast } from '../services/AirshipInstance'
 import { type ThemeProps, withTheme } from '../services/ThemeContext.js'
 import { EdgeText } from '../themed/EdgeText'
 import { MainButton } from '../themed/MainButton.js'
-import { Tile } from '../themed/Tile'
 
 type LocalState = {
   showAddBundledTxs: boolean,
