+++ src/components/scenes/FioDomainRegisterScene.js
@@ -11,6 +11,7 @@ import s from '../../locales/strings.js'
 import { connect } from '../../types/reactRedux.js'
 import { type NavigationProp } from '../../types/routerTypes.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
+import { Tile } from '../common/tiles/Tile'
 import { TextInputModal } from '../modals/TextInputModal.js'
 import type { WalletListResult } from '../modals/WalletListModal'
 import { WalletListModal } from '../modals/WalletListModal'
@@ -21,7 +22,6 @@ import { EdgeText } from '../themed/EdgeText'
 import { FormError } from '../themed/FormError'
 import { MainButton } from '../themed/MainButton.js'
 import { SceneHeader } from '../themed/SceneHeader'
-import { Tile } from '../themed/Tile'
 
 type LocalState = {
   selectedWallet: EdgeCurrencyWallet | null,
