+++ src/components/scenes/FioDomainSettingsScene.js
@@ -12,13 +12,13 @@ import { getDomainSetVisibilityFee, getRenewalFee, getTransferFee, renewFioDomai
 import { connect } from '../../types/reactRedux.js'
 import { type NavigationProp, type RouteProp } from '../../types/routerTypes.js'
 import { SceneWrapper } from '../common/SceneWrapper'
+import { Tile } from '../common/tiles/Tile'
 import { ButtonsModal } from '../modals/ButtonsModal'
 import { Airship, showError } from '../services/AirshipInstance'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext'
 import { ClickableText } from '../themed/ClickableText.js'
 import { EdgeText } from '../themed/EdgeText'
 import { MainButton } from '../themed/MainButton.js'
-import { Tile } from '../themed/Tile'
 
 type State = {
   showRenew: boolean,
