+++ src/components/scenes/RequestScene.js
@@ -23,6 +23,7 @@ import type { GuiCurrencyInfo, GuiDenomination } from '../../types/types.js'
 import { getAvailableBalance, getWalletName } from '../../util/CurrencyWalletHelpers.js'
 import { convertNativeToDenomination, getCurrencyInfo, getDenomFromIsoCode, getObjectDiff, truncateDecimals } from '../../util/utils.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
+import { FiatText } from '../common/text/FiatText.js'
 import { ButtonsModal } from '../modals/ButtonsModal.js'
 import { QrModal } from '../modals/QrModal.js'
 import { type WalletListResult, WalletListModal } from '../modals/WalletListModal.js'
@@ -31,7 +32,6 @@ import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services
 import { Card } from '../themed/Card.js'
 import { EdgeText } from '../themed/EdgeText.js'
 import { type ExchangedFlipInputAmounts, ExchangedFlipInput } from '../themed/ExchangedFlipInput.js'
-import { FiatText } from '../themed/FiatText.js'
 import { FlipInput } from '../themed/FlipInput.js'
 import { QrCode } from '../themed/QrCode'
 import { ShareButtons } from '../themed/ShareButtons.js'
