+++ src/components/scenes/SendScene.js
@@ -28,6 +28,10 @@ import { type GuiExchangeRates, type GuiMakeSpendInfo } from '../../types/types.
 import { getWalletName } from '../../util/CurrencyWalletHelpers.js'
 import { convertTransactionFeeToDisplayFee } from '../../util/utils.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
+import { AddressTile } from '../common/tiles/AddressTile.js'
+import { EditableAmountTile } from '../common/tiles/EditableAmountTile'
+import { ErrorTile } from '../common/tiles/ErrorTile'
+import { Tile } from '../common/tiles/Tile.js'
 import { ButtonsModal } from '../modals/ButtonsModal'
 import { FlipInputModal } from '../modals/FlipInputModal.js'
 import { TextInputModal } from '../modals/TextInputModal.js'
@@ -35,13 +39,9 @@ import type { WalletListResult } from '../modals/WalletListModal'
 import { WalletListModal } from '../modals/WalletListModal'
 import { Airship, showError } from '../services/AirshipInstance.js'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
-import { AddressTile } from '../themed/AddressTile.js'
 import { EdgeText } from '../themed/EdgeText'
-import { EditableAmountTile } from '../themed/EditableAmountTile'
-import { ErrorTile } from '../themed/ErrorTile'
 import { PinDots } from '../themed/PinDots.js'
 import { SelectFioAddress } from '../themed/SelectFioAddress.js'
-import { Tile } from '../themed/Tile.js'
 
 const PIN_MAX_LENGTH = 4
