+++ src/components/scenes/Staking/StakeModifyScene.js
@@ -18,16 +18,16 @@ import { getPolicyIconUris, getPolicyTitleName, getPositionAllocations, stakePlu
 import { zeroString } from '../../../util/utils.js'
 import { FillLoader } from '../../common/FillLoader.js'
 import { SceneWrapper } from '../../common/SceneWrapper.js'
+import { CryptoFiatAmountTile } from '../../common/tiles/CryptoFiatAmountTile.js'
+import { EditableAmountTile } from '../../common/tiles/EditableAmountTile.js'
+import { ErrorTile } from '../../common/tiles/ErrorTile.js'
+import { IconTile } from '../../common/tiles/IconTile'
 import { FlipInputModal } from '../../modals/FlipInputModal.js'
 import { FlashNotification } from '../../navigation/FlashNotification.js'
 import { Airship } from '../../services/AirshipInstance.js'
 import { cacheStyles, useTheme } from '../../services/ThemeContext.js'
 import { Alert } from '../../themed/Alert.js'
-import { CryptoFiatAmountTile } from '../../themed/CryptoFiatAmountTile.js'
 import { EdgeText } from '../../themed/EdgeText.js'
-import { EditableAmountTile } from '../../themed/EditableAmountTile.js'
-import { ErrorTile } from '../../themed/ErrorTile.js'
-import { IconTile } from '../../themed/IconTile'
 import { SceneHeader } from '../../themed/SceneHeader.js'
 
 type Props = {
