+++ src/components/scenes/Staking/StakeOverviewScene.js
@@ -17,8 +17,8 @@ import { getWalletFiat } from '../../../util/CurrencyWalletHelpers.js'
 import { getAllocationLocktimeMessage, getPolicyIconUris, getPolicyTitleName, getPositionAllocations, stakePlugin } from '../../../util/stakeUtils.js'
 import { FillLoader } from '../../common/FillLoader'
 import { SceneWrapper } from '../../common/SceneWrapper.js'
+import { CryptoFiatAmountTile } from '../../common/tiles/CryptoFiatAmountTile.js'
 import { cacheStyles, useTheme } from '../../services/ThemeContext.js'
-import { CryptoFiatAmountTile } from '../../themed/CryptoFiatAmountTile.js'
 import { MainButton } from '../../themed/MainButton.js'
 import { SceneHeader } from '../../themed/SceneHeader.js'
 import { StakingReturnsCard } from '../../themed/StakingReturnsCard.js'
