+++ src/components/scenes/TransactionDetailsScene.js
@@ -29,6 +29,7 @@ import {
   truncateDecimals
 } from '../../util/utils.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
+import { Tile } from '../common/tiles/Tile.js'
 import { AccelerateTxModel } from '../modals/AccelerateTxModel.js'
 import { ContactListModal } from '../modals/ContactListModal.js'
 import { RawTextModal } from '../modals/RawTextModal.js'
@@ -39,7 +40,6 @@ import { Airship, showError } from '../services/AirshipInstance.js'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 import { EdgeText } from '../themed/EdgeText.js'
 import { MainButton } from '../themed/MainButton.js'
-import { Tile } from '../themed/Tile.js'
 
 type OwnProps = {
   route: RouteProp<'transactionDetails'>
