+++ src/components/themed/ControlPanel.js
@@ -29,13 +29,13 @@ import { useDispatch, useSelector } from '../../types/reactRedux'
 import { type NavigationProp, type ParamList, Actions } from '../../types/routerTypes.js'
 import { getWalletFiat } from '../../util/CurrencyWalletHelpers.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
+import { FiatText } from '../common/text/FiatText'
 import { ButtonsModal } from '../modals/ButtonsModal.js'
 import { ScanModal } from '../modals/ScanModal'
 import { type WalletListResult, WalletListModal } from '../modals/WalletListModal.js'
 import { Airship, showError } from '../services/AirshipInstance.js'
 import { type Theme, cacheStyles, useTheme } from '../services/ThemeContext'
 import { DividerLine } from './DividerLine'
-import { FiatText } from './FiatText.js'
 import { TitleText } from './TitleText'
 
 type Props = { navigation: NavigationProp<'controlPanel'> }
