+++ src/modules/FioAddress/components/FioActionSubmit.js
@@ -5,6 +5,7 @@ import type { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 import { ActivityIndicator, View } from 'react-native'
 
+import { Tile } from '../../../components/common/tiles/Tile'
 import type { WalletListResult } from '../../../components/modals/WalletListModal'
 import { WalletListModal } from '../../../components/modals/WalletListModal'
 import { Airship, showError, showToast } from '../../../components/services/AirshipInstance'
@@ -12,7 +13,6 @@ import type { Theme, ThemeProps } from '../../../components/services/ThemeContex
 import { cacheStyles, withTheme } from '../../../components/services/ThemeContext'
 import { EdgeText } from '../../../components/themed/EdgeText'
 import { MainButton } from '../../../components/themed/MainButton.js'
-import { Tile } from '../../../components/themed/Tile'
 import { FIO_STR } from '../../../constants/WalletAndCurrencyConstants'
 import s from '../../../locales/strings'
 import { getDisplayDenomination } from '../../../selectors/DenominationSelectors.js'
